apiAI/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
motor
python-dotenv
joblib